msm: gsi: get rid of TERR(), TDBG() and PRT_STAT()
The TERR() and TDBG() macros in "gsi_dbg.c" are defined in a way
that obscures the fact that they amount to calls to pr_err() and
pr_dbg().  What they do is normally accomplished by defining the
pr_fmt() macro at the top of the source file.  So define pr_fmt(),
and replace calls to TERR() and TDBG() to use pr_err() and pr_dbg()
instead.

PRT_STAT() is unlike the other two, in that it does not insert the
function name and source file line number in its output.  So we
can't just use pr_err()  now that we've defined pr_fmt().  In this
case, translate all PRT_STAT() into what they really amount to,
which is a call to printk() with a KERN_ERR character prepended.

Change-Id: Iafbd66f3773eaeccd25aec88c771a0dc56c39a30
Signed-off-by: Alex Elder <elder@codeaurora.org>