msm: gsi: simplify gsi_debugfs_init()
It doesn't really matter *which* file or directory fails to be
created when setting up the debugfs hierarchy for GSI.  If any
failure occurs inside gsi_debugfs_init(), *nothing* in the hierarchy
will exist on return.  Just use a common report that an error
occurred, and get rid of the all the pr_err() calls that provide the
more detailed information.

Change-Id: I4b58dbe717982c7f7bc060a8add007976203bd75
Signed-off-by: Alex Elder <elder@codeaurora.org>