msm: gsi: use "msm_gsi.h" definitions
Remove the definitions that are defined redundantly in the GSI
private header "gsi.h", and use the version defined in the public
"include/linux/msm_gsi.h" instead.  This way things are defined
exactly once, and we are assured code inside and outside the GSI
source uses the same definitions.

Change-Id: Iaab54fb5245127c40bd4362fee9e52339eb84572
Signed-off-by: Alex Elder <elder@codeaurora.org>