msm: gsi: get rid of request and release clock callbacks
The GSI context structure allows a clock request and clock release
callback to be defined at GSI device registration time.  If defined,
they are used around interrupt handling code--calling the request
callback before handling the interrupt and then the release callback
after interrupt handling is complete.

But gsi_register_device() is only called in one spot, and that case
does not provide clock request and release callback functions.  So
this functionality is unused at this time.

Delete these callback function pointers from the GSI context
structure, and delete the code that uses it, to simplify the
interrupt handler.

Change-Id: I81dbf371ca024652c37f0045352630480f51a560
Signed-off-by: Alex Elder <elder@codeaurora.org>