msm: gsi: don't request event channel id
Since we only support IRQ interrupts, we never need to supply a
desired event channel id when allocating an event ring.  Get rid of
the "evchid" and "evchid_valid" fields in the gsi_evt_ring_props
structure, as well as the code that depends on whether an id was
supplied at event channel allocation time.

Change-Id: I4274588a348f9e4d6578375b8b4faf1d901cfbcb
Signed-off-by: Alex Elder <elder@codeaurora.org>