msm: gsi: MSI is not used for event rings
Since we only support IRQ interrupts, we don't need to program
anything special for MSI interrupts for an event ring.  We already
are programming 0 for the "intvec" value, and 0 for the high and low
MSI address for each ring.  Make that fact more obvious by doing it
explicitly, and eliminate the "intvec" and "msi_addr" fields from
the event ring properties structure.

Change-Id: I37ac79793b24ba543aa63821958b51366d84693c
Signed-off-by: Alex Elder <elder@codeaurora.org>