msm: gsi: we don't use the read pointer update address
The rp_update_addr field in an event ring properties structure
is meant to hold the physical address to which the ring's read
pointer is written whenever an event is generated.  This feature
may be used at a future date, but for right now it is unused.  We
indicate this by writing 0's when setting up the event ring.  We
don't need the "rp_update_addr" field, so get rid of it.

Change-Id: I1948fb2e7b90b46787110a14336f7a1d50556397
Signed-off-by: Alex Elder <elder@codeaurora.org>