msm: gsi: don't call devm_free_irq() in gsi_register_device()
An IRQ in requested in gsi_register_device() using the managed
device resource version of requesting an IRQ, devm_request_irq().
As such the IRQ will automatically be freed when the device
structure detached, so there is no need to call devm_free_irq()
if an error occurs.  There is also no reason to call that in
gsi_deregister_device().

Change-Id: If29a70ea6dfcf009e45653a032ac5364511707ce
Signed-off-by: Alex Elder <elder@codeaurora.org>