msm: gsi: use ETIMEDOUT in gsi_alloc_evt_ring()
This patch is the first in a series that aims to use standard Linux
errno values in place of GSI_STATUS_*.  The GSI_STATUS_* values are
in the range [0..12], so we need to be a little careful to ensure we
don't use a standard errno value in that range as we make this
conversion.  We'll start with replacing GSI_STATUS_TIMED_OUT (= 11)
with ETIMEDOUT (= 60).

These return values are passed up a chain of called functions, so we
also need to be careful to ensure we don't break anything that might
not handle the new return value properly.  Most of the explanation
in this and the following patches amounts to demonstrating that
changing the return value does not change the behavior of the code.

In gsi_alloc_evt_ring(), if wait_for_completion_timeout() returns 0
it indicates a timeout occurred, and in this case the function
returns -GSI_STATUS_TIMED_OUT.  Return -ETIMEDOUT instead.

There are three callers of gsi_alloc_evt_ring():
  - ipa3_request_gsi_channel() only checks whether GSI_STATUS_SUCCESS
    is returned, and if not, it passes to its caller -EFAULT.  So this
    site is unaffected by using the new -ETIMEDOUT value.
  - ipa3_alloc_common_event_ring() checks to see if a non-zero value
    is returned, and if so, returns that result to its caller.
      - ipa3_alloc_common_event_ring() is called only by
        ipa3_setup_apps_pipes(), and that checks for (and otherwise
        ignores) a no-zero return value.  So this site is not affected
        by the change to return -ETIMEDOUT.
  - ipa_gsi_setup_channel() checks to see if a non-zero value is
    returned, and if so, returns that result to its caller.
      - ipa_gsi_setup_channel() is called by ipa3_setup_sys_pipe(),
        which checks to see if a non-zero value is returned, and if
        so, returns that result to its caller.
          - ipa3_setup_sys_pipe() is called by ipa3_setup_apps_pipes(),
            twice.  In both of these spots, all that matters is whether
            non-zero value gets returned so these sites are not affected
            by the change to return -ETIMEDOUT.
          - ipa3_setup_sys_pipe() is called by handle3_ingress_format()
            and handle3_egress_format() also.  Both of these translate
            any returned error to -EFAULT.  So neither of these spots
            is affected by the change to return -ETIMEDOUT.

So in summary, all of the paths that might receive a timeout
indication that originates from gsi_alloc_evt_ring() will exhibit
the same behavior whether that function returns GSI_STATUS_TIMED_OUT
or ETIMEDOUT.

Change-Id: I9e6ef4d7ba2d51fddc1be388f0fb8a4553fa1584
Signed-off-by: Alex Elder <elder@codeaurora.org>