msm: gsi: use EAGAIN rather than GSI_STATUS_AGAIN
This patch replaces the use of GSI_STATUS_AGAIN (= 11) with the
standard Linux errno EAGAIN (also = 11).  Because they have the
same numeric value, there will be no difference in the behavior of
the code when making the change, so we don't need to do anything
special to verify code paths.  With no remaining references to
GSI_STATUS_AGAIN, we can remove its definition.

Change-Id: Ib63af9420da3f59061b5c95666d24cd53e30655c
Signed-off-by: Alex Elder <elder@codeaurora.org>