msm: gsi: use ENODEV in gsi_register_device()
Use ENODEV (= 19) rather than GSI_STATUS_NODEV (= 8) in
gsi_register_device().  The only caller is ipa3_post_init(),
which returns -ENODEV to its caller if it gets anything other
than GSI_STATUS_SUCCESS.

Make the same change in gsi_deregister_device(), which is
called also (and only) by ipa3_post_init() in an error path.
In this case the returned value is ignored.

In both cases the code will work the same regardless of the
particular value returned to indicate the GSI device is not
properly initialized.

Change-Id: I504bebd2d3e7c97e0d93f20cd55a40a4418f4a7e
Signed-off-by: Alex Elder <elder@codeaurora.org>