msm: gsi: use ENOSPC rather than GSI_STATUS_RING_INSUFFICIENT_SPACE;
GSI_STATUS_RING_INSUFFICIENT_SPACE is used in only one spot, as a
return value gsi_queue_xfer() in the event a transfer is requested
with more transfers than available slots.  Return the standard Linux
ENOSPC in this location instead.

An earlier patch explained how the six callers of gsi_queue_xfer()
handle returned error codes the same way regardless of their
partical values.  So this change has no effect on the behavior of
the code.

Change-Id: Ib064019ad29b631a4ef7f6169b379d8cbf9dea7b
Signed-off-by: Alex Elder <elder@codeaurora.org>