msm: gsi: return a negative value to indicate poll empty
When gsi_poll_channel() finds that there is no new data, it returns
GSI_STATUS_POLL_EMPTY to its caller.  All other return values are
negative codes (or GSI_STATUS_SUCCESS).  Change the code to return a
negative value in this case.  -GSI_STATUS_POLL_EMPTY

gsi_poll_channel() is only called by ipa_poll_gsi_pkt(),
and if it returns GSI_STATUS_POLL_EMPTY (or now, negative)
it returns that value to its caller.

ipa_poll_gsi_pkt() is called in a loop by ipa_poll_gsi_pkt() and
ipa3_handle_rx_core().  In both of these, if a non-zero value is
returned, the loop is exited and count is returned (independent of
the non-zero return value).  The only thing that matters is whether
the returned value is zero or not.  So the behavior will be the same
whether the return vale is negative or not.

Change-Id: Ie957d3ddbb80510e308b63d19943221e66b0b3bb
Signed-off-by: Alex Elder <elder@codeaurora.org>