msm: gsi: simplify ipa3_debugfs_init()
It doesn't really matter *which* file or directory fails to be
created when setting up the debugfs hierarchy for IPA.  If any
failure occurs inside ipa3_debugfs_init(), *nothing* in the
hierarchy will exist on return.  Just use a common report that an
error occurred, and get rid of the all the IPAERR() calls that
provide the more detailed information.

Change-Id: I3d585a4d5a2bd2ff84a65c5b03bd5ce9ff710a5e
Signed-off-by: Alex Elder <elder@codeaurora.org>