msm: ipa: have gsi_register_device() return a handle
Change the prototype for gsi_register_device() so it returns a
device handle pointer rather than an integer error code.  This is a
much more common way of handling this sort of thing than passing in
the address of an object to receive the result.  If an error occurs,
return a pointed-coded error value.

Change-Id: I6cf8dfb26ae7573baa06560da0f8e49321577e57
Signed-off-by: Alex Elder <elder@codeaurora.org>