msm: ipa: always use GSI_NO_EVT_ERINDEX to represent no event ring
A special value ~0 is used in some cases to represent "no event
ring handle," and it's passed in the properties structure to
gsi_alloc_channel().  In that function, if ~0 is passed, the handle
value used is converted to GSI_NO_EVT_ERINDEX.

There's no point in using two values to represent the same thing.
Just use GSI_NO_EVT_ERINDEX always, and stop using the magic ~0
value.  It's clearer, and ends up simplifying a little code.

Change-Id: Ie46a117fd8cfbffe214a60a7a4060441b6bfb318
Signed-off-by: Alex Elder <elder@codeaurora.org>