msm: gsi: git rid of needless checks in gsi_deregister_device()
There is only one caller of gsi_deregister_device(): it is called in
the cleanup path for ipa3_post_init().  It is only called after
gsi_register_device() is known to have succeeded.  In that case, the
device handle passed is ipa3_ctx->gsi_dev_hdl, which is the value
returned by gsi_register_device().  And when successful, that value
is gsi_ctx.

We know, therefore, that gsi_ctx will be non-null (it must have been
for gsi_register_device() to succeed).  We also know that the only
place gsi_ctx->per_registered is set is in gsi_register_device()
when it is successful.  And we know that ipa3_ctx->gsi_dev_hdl will
equal gsi_ctx, because it has just been set to that value.

Three tests at the top of gsi_deregister_device() can therefore be
eliminated, because the conditions they check will never occur.

Change-Id: I8b7c0146eb40a47036f448dfe54afee9b3eced5c
Signed-off-by: Alex Elder <elder@codeaurora.org>