msm: gsi: get rid of GSI_GSI_REG_BASE_OFFS
GSI_GSI_REG_BASE_OFFS is defined at the top of "gsi_reg.h" with the
value 0, and then every other register offset defined within the
file is defined in terms of that.

There is really no reason to have a global offset like this.  If all
addresses are shifted by a constant amount, then the address range
that is mapped for those addresses can just be shifted.

The value is zero anyway, so simply delete the definition and stop
adding it to all the register offsets definitions.  This patch makes
that simple change; the next patch will clean up the result a little
more.

Change-Id: I749034ebe76c6d46c8adc2213ed9c9817aea34ca
Signed-off-by: Alex Elder <elder@codeaurora.org>