msm: gsi: be more obvious about programming ring base address
The 64-bit ring base address (for an event ring or a channel ring)
is progammed by storing the address's top half and bottom half into
two context registers for the event or channel (CNTXT_2 and CNTXT_3).

It is not easy to see that this simple thing is what's going on when
looking at the code that performs the writes.  We know that the
entire upper and lower 32 bits are used in these two writes; it's
actually clearer to see the masking and shifting get done directly
rather than using the MASK and SHIFT values otherwise defined in
"gsi_reg.h".

Change-Id: I4f00671fad10d320c8deb56741ab49aa5d72b726
Signed-off-by: Alex Elder <elder@codeaurora.org>