msm: gsi: be more obvious about programming doorbell address
The 64-bit doorbell address (for an event ring or a channel ring)
is progammed by storing the address's top half and bottom half
into two context registers for the event or channel (DOORBELL_0
and DOORBELL_1).

It is not easy to see that this simple thing is what's going on
when looking at the code that performs the writes.  We know that
the entire upper and lower 32 bits are used in these two writes;
it's actually clearer to see the masking and shifting get done
directly rather than using the MASK and SHIFT values otherwise
defined in "gsi_reg.h".

Change-Id: Iff200f9929888a4918554d582988b08048e644c3
Signed-off-by: Alex Elder <elder@codeaurora.org>