msm: gsi: eliminate *_BMSK symbols for 32-bit fields
The definitions in "gsi_reg.h" assume that all fields can be of any
width and at any position within a register.  In a lot of cases,
there is only one field in a given register, and that field consumes
the entire 32 bits.  In those cases, the *_BMSK symbol (such as
GSI_GSI_MANAGER_MCS_CODE_VER_VER_BMSK) is never used in code, nor is
the associated *_SHFT register (which is always 0, for example
GSI_GSI_MANAGER_MCS_CODE_VER_VER_SHFT).

Delete the definitions of the *_BMSK and *_SHFT symbols for
registers that contain a single 32-bit field.

Change-Id: I75f8de962dc1aaca4d5eb6c582cdae2689d75976
Signed-off-by: Alex Elder <elder@codeaurora.org>