msm: gsi: don't continue if GSI is not enabled
There is a check fairly late in gsi_register_device() to see whether
the basic GSI configuration has been completed, checking the ENABLED
bit in the STATUS register for the AP execution environment (EE).
If it is not set, it issues a warning but proceeds as if this were
not an error.

But it *is* an error, and as stated in the warning message the GSI
is not usable in this case.

If this condition is seen, have gsi_register_device() return an
error so we stop any further IPA initialization.  Furthermore, do
this earlier than before--right after we have mapped the GSI
memory--so subsequent reads and writes to GSI memory are known to be
working with initialized hardware.

Change-Id: I611a92cfe50861f260253d8810f500b338988178
Signed-off-by: Alex Elder <elder@codeaurora.org>