msm: gsi: introduce gsi_irq_set()
If the mask value provided to gsi_irq_update() is ~0, none of the
previous values saved in the IRQ mask register will be preserved.
So there's no point in doing the read-modify-write; just do a write
of the register instead.

Create gsi_irq_set() to abstract this notion, and use it in all the
places that currently pass ~0 as the mask for gsi_irq_update().

Change-Id: Ief7353ef8d28c28e52c28d1c282cc8eed0d50f70
Signed-off-by: Alex Elder <elder@codeaurora.org>