msm: ipa: eliminate qcom,ipa-pipe-mem DT property
The ipa_pipe_mem_start_ofst and ipa_pipe_mem_size fields of the
ipa3_plat_drv_res structure are assigned but never used in the IPA
lite code.  Their values are assigned based on an optional Device
Tree property "qcom,ipa-pipe-mem", which is never defined for the
SDM845.

Delete the unused struct fields.  This makes the code that looks up
the value of "qcom,ipa-pipe-mem" unnecessary, so remove that and
delete that property's definition from the DT documentation.

Change-Id: I53f11820d2577a5f60c10c59855221b6980c5e6c
Signed-off-by: Alex Elder <elder@codeaurora.org>