msm: ipa: eliminate qcom,ipa-hw-mode DT property
Both the ipa3_context and ipa3_plat_drv_res structures have a field
named "ipa_hw_mode" to record the memory allocation mode for the IPA
hardware (virtual memory, memory behind a PCIE bridge, or "normal").
By default, the mode is normal, but virtual or PCIE mode can be
indicated in the value of the "qcom,ipa-hw-mode" Device Tree
property.  For the the SDM845, the "sdm845.dtsi" has value 1 for
this, which represents virtual mode.

This is fine, but nothing present in the IPA lite code base ever
references the "ipa_hw_mode" field once it's assigned.  So the value
in the DT file has no effect whatsoever.

So delete this property, and the code that supports it.  Get rid of
the property in "sdm845.dtsi" as well.

Change-Id: I22dfebda14f4693032bf1ab13ac1566cd7c0679b
Signed-off-by: Alex Elder <elder@codeaurora.org>