msm: ipa: eliminate qcom,use-ipa-tethering-bridge DT property
The ipa3_context and ipa3_plat_drv_res structures have a Boolean
field named "use_ipa_teth_bridge" to record whether a tethering
bridge driver should be used.  It is set to true if the DT
contains the "qcom,use-ipa-tethering-bridge".  For the the SDM845,
"sdm845.dtsi" does contain this property.

Nothing present in the IPA lite code base ever references the
"use_ipa_teth_bridge" field once it's assigned.  So whether the
value is present in the DTB has no effect whatsoever.

So delete this property, and the code that supports it.  Delete the
property from "sdm845.dtsi" as well.

Change-Id: I8de392566c0138b20492a655c406b016989af8ea
Signed-off-by: Alex Elder <elder@codeaurora.org>