msm: ipa: eliminate ipa3_get_gsi_ver()
We know that the IPA hardware version is 3.5.1 (IPA_HW_v3_5_1).
This is verified in get_ipa_dts_configuration(), which fails if the
DT indicates anything else.

Given that, we know ipa3_get_gsi_ver() will indicate that the GSI
hardware version is 1.3.0 (GSI_VER_1_3).  Just hard-code that value
and eliminate ipa3_get_gsi_ver().

Change-Id: Icd48d7af76ae2d0a677a629f69a4a4d8fb9ef1ab
Signed-off-by: Alex Elder <elder@codeaurora.org>