msm: ipa: don't create "hw_type" debugfs file
There is a debugfs file "hw_type" that shows just "13" to identify
the version of IPA that's in use.  We only support the SDM845, which
has version 3.5.1 of IPA hardware, so "hw_type" does not provide any
information we don't already know.  Get rid of it (with a plan to
put something like it back again when we support other versions).

Change-Id: I1f2e668b020bf77afc4414d4edec1f3c28cd427e
Signed-off-by: Alex Elder <elder@codeaurora.org>