msm: ipa: eliminate ipa_hw_type field from ipa3_plat_drv_res
There remain no users of the ipa3_plat_drv_res field the platform
driver resource structure.  The only reason we need that information
is at initialization time, where we verify the DTB states we've got
the version of IPA hardware we support (3.5.1).  Fetch that
information into a local variable, and don't bother recording it in
the resource structure any more.

Change-Id: Ic322e4552e44b115fad88c24e1efde676ba7ef05
Signed-off-by: Alex Elder <elder@codeaurora.org>