msm: ipa: use "gsi" instead of "transport" in names
Several fields in the IPA context structure begin with "transport".
This is a generic term that may be more appropriate for the possible
ways IPA can be used, but four our purpose, the transport interface
is the GSI interface.  This is a little more obvious when the fields
are named things like "gsi_irq" rather than "transport_irq".

Change-Id: Id40141427102b92e9f933e5a1db6826d991c9cc2
Signed-off-by: Alex Elder <elder@codeaurora.org>