msm: ipa: eliminate qcom,ipa-wdi2 DT property
The ipa3_context and ipa3_plat_drv_res structures have a Boolean
field named "ipa_wdi2 to record whether the WDI 2.0 is in use.  It
is set to true if the DT contains the "qcom,ipa-wdi2".  For the the
SDM845, "sdm845.dtsi" contains this property.

The flag "ipa_wdi2 flag is never used in the IPA lite code after
it's set.  So delete the field, delete the Device Tree property, and
remove the property from "sdm845.dtsi".

Change-Id: Id1918ab90a26d6c1a4c18d9e343862503bbb53cf
Signed-off-by: Alex Elder <elder@codeaurora.org>