msm: ipa: eliminate qcom,arm-smmu and qcom,msm-smmu DT properties
The IPA code supports DT properties that indicate whether an
SMMU is used, and if so, which variety is used.  The code that
parses this information makes it clear that only one of these cases
can hold: ARM; MSM; or no SMMU.

The SDM845 defines "qcom,arm-smmu" in its "sdm845.dtsi".  We can
delete a big block of conditional code by supporting only what the
SDM845 uses, and we can therefore remove the two SMMU properties
that indicate which is used from the Device Tree.

We can also remove the arm_smmu field of the smmu_info structure,
because it's never used after being set.

Change-Id: I934c1fadec8de84d6539922d39eb45819970450a
Signed-off-by: Alex Elder <elder@codeaurora.org>