msm: ipa: eliminate platform use_64_bit_dma_mask field
In get_ipa_dts_configuration(), the DTB is consulted to determine
whether the Boolean property "qcom,use-64-bit-dma-mask" is present.
The result is stored in the use_64_bit_dma_mask field of the
ipa3_plat_drv_res (first) and ipa3_context (later) structures.
But ultimately this has no effect on the IPA platform code.

Instead, the only place that property has an effect is when it is
recorded in the smmu_info structure, which also has a Boolean field
with the same name, whose value is assigned in exactly the same way.

Delete th use_64_bit_dma_mask fields from both the ipa3_plat_drv_res
and ipa3_context structures, and get rid of the code that uses them.

Change-Id: Icf5c2975ded85b123dad418adc8192a6aa651ce1
Signed-off-by: Alex Elder <elder@codeaurora.org>