msm: ipa: eliminate qcom,use-64-bit-dma-mask DT property
The smmu_info structure has a Boolean field "use_64_bit_dma_mask"
to record whether the DMA to use for IPA should be 64 bits (or 32).
If the "qcom,use-64-bit-dma-mask" property is present in the DTB,
a 64-bit mask is used.  For SDM845, "sdm845.dtsi" does contain this
property.

Change the code so it uses a 64-bit DMA mask unconditionally, and
get rid of the Boolean flag and the DT property that makes this be
optional behavior.

Change-Id: I95e930c31cb9705ec8a94ed795b9064ba44faf4c
Signed-off-by: Alex Elder <elder@codeaurora.org>