msm: ipa: eliminate qcom,bandwidth-vote-for-ipa DT property
The ipa3_plat_drv_res structure has a Boolean field "use_bw_vote" to
determine whether the bus scaling driver should be used for
controlling the core clock for IPA.  The value of that field is set
based on whether the "qcom,bandwidth-vote-for-ipa" DT property is
present.

That property is present for SDM845 in "sdm845.dtsi".  Simplify the
code by assuming this, and get rid of the Boolean field and the DT
property.

Change-Id: I37dbaa1b4d1fa3a6a82b9758517f64b543c3cd6b
Signed-off-by: Alex Elder <elder@codeaurora.org>