msm: ipa: eliminate IPA RX ring size DT properties
The ipa3_plat_drv_res structure has two 32-bit fields indicating the
size of the RX ring to use for the WAN and LAN.  They're specified
in the Device Tree using properties "qcom,wan-rx-ring-size" and
"qcom,lan-rx-ring-size", respectively, and they take on a default
value (192 each) if not specified in the DTB.

Neither of these is specified for the SDM845 in "sdm845.dtsi", so
they take on their default values.  Simplify the code by assuming
this, and get rid of the fields in the ipa3_plat_drv_res structure
that represent them.  Also get rid of the DT properties that allow
them to be specified, because only the default is supported.

Change-Id: Ibd530f481ebf95b163f35b96337f944097e90fb4
Signed-off-by: Alex Elder <elder@codeaurora.org>