msm: ipa: eliminate IPA context RX ring size fields
The IPA context contains two fields to indicate the size of the LAN
and WAN RX rigns: lan_rx_ring_size and wan_rx_ring_size.  These
fields are only set to a constant value.  Simplify the code by using
that constant value in the only places these fields are used, and
eliminate the fields.

Change-Id: Ie7db9ef735a88af486c5b4394baf3fa4de0b1bb4
Signed-off-by: Alex Elder <elder@codeaurora.org>