msm: ipa: eliminate qcom,skip-uc-pipe-reset DT property
In get_ipa_dts_configuration(), the DTB is consulted to determine
whether the Boolean property "qcom,skip-uc-pipe-reset" is present.
The result is stored in the skip_uc_pipe_reset field of the
ipa3_plat_drv_res (first) and ipa3_context (later) structures.
Ultimately this has no effect though, in that it's never used after
being set in the IPA lite code.  So delete the property and the code
that supports it.

For SDM845, "sdm845.dtsi" does not contain this property.

Change-Id: I6db4179e3b97a10c5eae7f4a0a57ef534322b157
Signed-off-by: Alex Elder <elder@codeaurora.org>