msm: ipa: eliminate ipa3_ctx->apply_rg10_wa
The SDM845 does not require the register group 10 access workaround.
Delete all IPA code that's dependent on that being required, and
delete the Boolean flag in an IPA context that indicates that.

The result leaves a few things unused, or trivial, and the next few
patches will tidy those up.

Change-Id: I72591645128ed9b024485149d8a6311f4b994ca4
Signed-off-by: Alex Elder <elder@codeaurora.org>