msm: ipa: eliminate the qcom,ipa-tz-unlock-reg DT property
The "qcom,ipa-tz-unlock-reg" Device Tree property is intended to
define a set of address ranges that IPA needs to be unlocked by
Trust Zone.  It is not needed for the SDM845, and since that's our
only target platform, it's not needed.

Eliminating support for that property means the ipa_tz_unlock_reg
and ipa_tz_unlock_reg_num fields present in both the ipa3_ctx and
ipa3_plat_drv_res structures are no longer necessary.  And as a
result several blocks of code can be removed.

Change-Id: I45072412fc3d93e8128cc63db44ada4a997d662a
Signed-off-by: Alex Elder <elder@codeaurora.org>