msm: ipa: ipa3_ctx->smmu_s1_bypass is not used
The smmu_s1_bypass field of the ipa3_ctx structure is never used
after it's set.  So the field and the code associated with that
field can be removed.

With that, ipa3_ctx->smmu_present is no longer ever used after being
set, so that too can go away.

That makes the present field of the smmu_info structure unnecessary,
so it can be removed.

Change-Id: I79647e55afd2025465132c688d3012ce1173254d
Signed-off-by: Alex Elder <elder@codeaurora.org>