msm: ipa: duplicate ipa_gsi_notify_cb() in "gsi.c"
Currently, when registering a GSI device, a notify callback
function must be provided.  A GSI device is registered only
once, in ipa3_post_init(), and in that location the callback
function provided is ipa_gsi_notify_cb().

The only thing ipa_gsi_notify_cb() does is log a message that
indicates what event occurred (and in most cases, results in a BUG()
call).  There is really nothing IPA-related about what that function
does, so there's no real reason it has to be associated with the IPA.

Make a copy of the function in "gsi.c", rename it gsi_notify(),
and change the logging calls it makes to use GSIERR() rather than
IPAERR().  We aren't yet using this duplicate function, so add a
no-op reference to it in gsi_handle_glob_err() where it will first
be needed.

Change-Id: I662fdc0a94bff90ca24d445663ac64f9a94f070f
Signed-off-by: Alex Elder <elder@codeaurora.org>