msm: gsi: don't use gsi_notify() for global errors
The only time GSI_PER_EVT_GLOB_ERROR is used as the event id for a
call to gsi_notify() is in gsi_handle_glob_err(), when the event
type indicates a global (rather than channel) error.

When the event id is GSI_PER_EVT_GLOB_ERROR, gsi_notify() simply
logs two error messages to indicate the type of event and an error
descriptor.

Rather than calling gsi_notify() in this case, just open code what
it does in the single call site.  Change what's logged to indicate
the error type we got from the hardware.

GSI_PER_EVT_GLOB_ERROR is no longer used, so it can be deleted.

Change-Id: I4895c04a79f1957484b1743fa763cea8e8fb8c04
Signed-off-by: Alex Elder <elder@codeaurora.org>