msm: gsi: open-code global EE interrupt handling
When gsi_handle_glob_ee() calls it, the only thing gsi_notify() does
is report the type of error that was indicated and calls BUG().
Instead of calling gsi_notify(), just log the message directly.
Change the message to more directly state what has occurred.

This makes GSI_PER_EVT_GLOB_GP2 and GSI_PER_EVT_GLOB_GP3 unnecessary
and they can be deleted.

Delete the definition of GSI_PER_EVT_GLOB_GP1 and the code that
supports it as well.  It is never supplied to gsi_notify().

Change-Id: Ib9b1944dbc2fa903a4116ef301b24efb0e948d9b
Signed-off-by: Alex Elder <elder@codeaurora.org>