msm: gsi: eliminate gsi_notify()
The general interrupt handler is the only remaining caller of
gsi_notify().  It calls it for four indicated interrupt types:
microcontroller stack overflow; command FIFO overflow; bus error;
and break point.  For all of these, gsi_notify() simply logs an
error message, and for all but break point, this is followed by a
BUG() call.  Little value is provided by gsi_notify(), therefore,
and removing it makes this fact a little more clear.

So open-code the logging that was done by gsi_notify() in the places
its called in gsi_handle_general().  Update the messages to more
directly state what happened.  This allows gsi_notify() to be
removed.  And in doing this, there is no longer any need for the
gsi_per_notify structure nor the gsi_per_evt enumerated type.

Change-Id: I5b9c09786b7064ffae4c186f2292eb49c563b1ee
Signed-off-by: Alex Elder <elder@codeaurora.org>