msm: ipa: ipa3_res is a global, part 2
The first parameter to ipa3_post_init() is a pointer to a platform
driver resource structure.  But the only value ever passed for that
parameter is the address of ipa3_res, a global variable.

Make it more obvious that we're dealing with ipa3_res in
ipa3_post_init() by dropping the first paramenter and directly
operating on the global variable.

Change-Id: I5c17ac9396663350ea688f4e879f5fdaad65b177
Signed-off-by: Alex Elder <elder@codeaurora.org>