msm: ipa: ipa3_res is a global
The first parameter to ipa3_pre_init() is a pointer to a platform
driver resource structure.  But the only value ever passed for that
parameter is the address of ipa3_res, a global variable.

Make it more obvious that we're dealing with ipa3_res in
ipa3_pre_init() by dropping the first paramenter and directly
operating on the global variable.

Change-Id: I7cd8ca1f0d14e1a19efd5bc10c780f721f066dd1
Signed-off-by: Alex Elder <elder@codeaurora.org>