msm: ipa: clean up ipa3_write()
Simplify and clean up ipa3_write() a bit.  The main change is to
stop trying to look at whatever is getting written.  It is
completely ignored anyway, and checking for something that's too
big, and copying it in from user space simply adds opportunities for
irrelevant errors to occur.  Instead, if a 0-byte write comes in,
simply return.  Otherwise proceed with IPA firmware load without
bothering to process the write data provided.

In addition, there are a two trivial changes:
  - fix a comment to say "concurrent" rather than "consequent"
  - report the successful firmware load immediately

Change-Id: I0042646c7569f120d59889797ebbfa209a978ba4
Signed-off-by: Alex Elder <elder@codeaurora.org>