msm: ipa: use ipa3_pdev global
Just before platform_get_resource_byname() is called, the global
ipa3_pdev global is assigned to be the address of the IPA platform
device structure.  Since this is global, it can be used inside
platform_get_resource_byname() rather than passing the same value in
as an argument.  This makes it more obvious that we're using the
IPA platform device and not any other.

Change-Id: I80c68ccb0d5e664b338311cfb17c79bf1158c483
Signed-off-by: Alex Elder <elder@codeaurora.org>