msm: ipa: hide type of global ipa3_res
There's no need to expose the type of ipa3_res; it's used only
inside "ipa.c".  Furthermore, the type is used only to define that
global variable, so we can just define the global with that type and
not bother giving the type a name.  Finally, static objects are
always initialized with zeroes, so there's no need to do so
explicitly.

Change-Id: Ib6e0709d3e960f182bef608c9285baba562f52c7
Signed-off-by: Alex Elder <elder@codeaurora.org>