msm: ipa: encapsulate getting hardware version
The version of IPA hardware is recorded in the DTB, and is fetched
at the top of get_ipa_dts_configuration().  Move the code that
determines this information into its own function.  This is
preparing for the next patch, which will arrange to determine this
right away (so we can quit immediately if the IPA hardware is not
recognized, or is a version that's supported).

Change-Id: Ice0705c664649b955eefe680f020e03bd5d0a4b6
Signed-off-by: Alex Elder <elder@codeaurora.org>