msm: ipa: test for unsupported hardware right away
Check the version of the IPA hardware first thing in the platform
driver probe function.  If the hardware is not supported or not
recognized, there's no point in doing GSI initialization.

Assign the global ipa3_pdev to point to the IPA platform device
structure as soon as we have verified we are have supported IPA
hardware.  (This point in the probe function should only be reached
once, so there's no need to verify it's non-null before assigning it.)
Assign master_dev at that time as well.

Change-Id: I32b0704414fe6b7818255621561f1a70ecf6cf0f
Signed-off-by: Alex Elder <elder@codeaurora.org>