msm: gsi: gsi_ctx is non-null in gsi_register_device()
There is no need to check whether gsi_ctx is non-null in
gsi_register_device().  The proof of this statement is a little
long...

First, gsi_ctx is initialized in msm_gsi_init().  If that is
successful, gsi_ctx will have been initialized to a non-null value.

We know gsi_register_device() is called only by ipa3_post_init().
And ipa3_post_init() is called via workqueue when ipa3_write() is
called.  ipa3_write() is the write file operation for the IPA
character device, and that can only be called if the character
device gets set up successfully.  That occurs when ipa3_pre_init()
is successful, and that is only called by ipa_smmu_ap_cb_probe().

If the DTB is configured properly, ipa_smmu_ap_cb_probe() will only
be called when the top-level probe for a device matching "qcom,ipa"
is nearly complete.  That is, only when of_platform_populate()
is used to populate the "sub-devices" of the top-level IPA node
in the DTB will a node matching "qcom,ipa-smmu-ap-cb" be probed.
And we'll never reach that point unless msm_gsi_init() is
successful.

Therefore gsi_register_device() is only called if msm_gsi_init() is
successful, and if that's successful we know gsi_ctx will be
non-null.

Change-Id: I7914e867f3bcecd388afa3c72a5950c9b7b4e2c1
Signed-off-by: Alex Elder <elder@codeaurora.org>