msm: gsi: look up GSI IRQ at registration time
Have the GSI rather than IPA code look up the IRQ that's supposed to
be used for GSI.  Search the DTB for the "gsi-irq" property in
gsi_register_device(), and if successful, assign the result into the
properties structure (which is later stored into the GSI context).
Use "ret" rather than "res" to hold an integer return value.

Change-Id: I31c1c3be0fe0dd38e5cf963287d5f7dbd08783cb
Signed-off-by: Alex Elder <elder@codeaurora.org>