msm: ipa: initialize ipa3_ctx->ee in ipa3_pre_init()
The only thing ipa3_res.ee is used for is to initialize ipa3_ctx->ee.
Since its value is never otherwise used, it is safe to look it up in
the Device Tree only when it's needed.  And in that case, there's no
need to hold the value in the ipa3_res structure at all--it can just
be assigned directly into the ipa3_ctx structure.

This leaves no use for ipa3_res, so it can be removed.

Change-Id: I8b3d68054caf4fc1939bdb2f86f1e6859bed9610
Signed-off-by: Alex Elder <elder@codeaurora.org>