msm: gsi: don't record memory range
When gsi_register_device() is called, the memory range to use for
GSI is fetched from the DTB.  Once that range has been mapped, there
nothing in the code ever has any use for the range mapped.  So
there's no need to keep it around.

Using available local variables to hold this information allows us
to delete the phys_addr and size fields from the gsi_per_props
structure.

Change-Id: I09a995f478edafa548e0234067e866d366b78921
Signed-off-by: Alex Elder <elder@codeaurora.org>