msm: gsi: do not assume 32 channels or events
Ghanim Fodi pointed out that the limiting value of 32 used in two
locations iterating over bits in an IRQ mask should really use
GSI_EVT_RING_MAX.  I noticed that there was one other case of this
in the code, as well as spots two where GSI_CHAN_MAX should be used.

This patch implements Ghanim's general suggestion throughout the
"gsi.c".  I have done this in a separate patch because the result
changes the behavior of the code (because GSI_CHAN_MAX is 31 and
GSI_EVT_RING_MAX is 23).

Change-Id: Ia906ab879fd89c2c6aaba4dc680afc1dd81a8cf3
Signed-off-by: Alex Elder <elder@codeaurora.org>