msm: ipa: don't bother with an array of SMMU context buffers
There are only two entries in the smmu_cb[] array, and each of them
is used independent of the other.  Since there is no common handling
there's no real benefit to having them be grouped in an array.  Just
use two separate context buffers structures in their place.  This
obviates the need for the ipa_smmu_cb_type enumerated type, so it
can go away.

Change-Id: I28c0459ff356dc0a9c4b7dec5fc9572ecdf8f45c
Signed-off-by: Alex Elder <elder@codeaurora.org>